+++ book/listings/css/2/style.css
@@ -1,3 +1,8 @@
 button.text-button {
   color: red;
 }
+
+button.image-button {
+  color: blue;
+}
+
