+++ book/listings/css/4/main.rs
@@ -1,7 +1,5 @@
-use gdk::Display;
-use gtk::gdk;
 use gtk::prelude::*;
-use gtk::{Application, ApplicationWindow, Button, CssProvider, StyleContext};
+use gtk::{Application, ApplicationWindow, Button};
 
 fn main() {
     // Create a new application
@@ -10,33 +8,20 @@ fn main() {
         .build();
 
     // Connect to signals
-    app.connect_startup(|_| load_css());
     app.connect_activate(build_ui);
 
     // Run the application
     app.run();
 }
 
-fn load_css() {
-    // Load the css file and add it to the provider
-    let provider = CssProvider::new();
-    provider.load_from_data(include_bytes!("style.css"));
-
-    // Add the provider to the default screen
-    StyleContext::add_provider_for_display(
-        &Display::default().expect("Error initializing GTK CSS provider."),
-        &provider,
-        gtk::STYLE_PROVIDER_PRIORITY_APPLICATION,
-    );
-}
-
 fn build_ui(app: &Application) {
     // ANCHOR: buttons
     // Create buttons
-    let button_1 = Button::with_label("Press me!");
-    let button_2 = Button::with_label("Press me!");
+    let button_1 = Button::with_label("Destructive");
+    let button_2 = Button::with_label("Suggested");
 
-    button_1.add_css_class("button_1");
+    button_1.add_css_class("destructive-action");
+    button_2.add_css_class("suggested-action");
     // ANCHOR_END: buttons
 
     // Create `gtk_box` and add buttons
@@ -45,7 +30,7 @@ fn build_ui(app: &Application) {
         .margin_bottom(12)
         .margin_start(12)
         .margin_end(12)
-        .spacing(12)
+        .spacing(6)
         .build();
     gtk_box.append(&button_1);
     gtk_box.append(&button_2);
