--- book/listings/css/4/style.css
@@ -1,3 +1,0 @@
-button.button_1 {
-  color: red;
-}
