+++ book/listings/css/6/main.rs
@@ -1,7 +1,9 @@
 mod window;
 
+use gdk::Display;
+use gtk::gdk;
 use gtk::prelude::*;
-use gtk::Application;
+use gtk::{Application, CssProvider, StyleContext};
 
 use crate::window::Window;
 
@@ -12,10 +14,24 @@ fn main() {
         .build();
 
     // Connect to signals
+    app.connect_startup(|_| load_css());
     app.connect_activate(build_ui);
 
     // Run the application
     app.run();
+}
+
+fn load_css() {
+    // Load the css file and add it to the provider
+    let provider = CssProvider::new();
+    provider.load_from_data(include_bytes!("style.css"));
+
+    // Add the provider to the default screen
+    StyleContext::add_provider_for_display(
+        &Display::default().expect("Error initializing GTK CSS provider."),
+        &provider,
+        gtk::STYLE_PROVIDER_PRIORITY_APPLICATION,
+    );
 }
 
 fn build_ui(app: &Application) {
