+++ book/listings/css/7/style.css
@@ -1,6 +1,3 @@
-button.button_1:hover {
-  color: red;
-  background-color: yellow;
-  background-image: none;
-  transition: 1s;
+combobox box arrow {
+  color: red 
 }
