+++ book/listings/css/7/window/window.ui
@@ -3,15 +3,17 @@
   <template class="MyGtkAppWindow" parent="GtkApplicationWindow">
     <property name="title" translatable="yes">My GTK App</property>
     <child>
-      <object class="GtkButton">
-        <property name="label">Hover over me!</property>
+      <object class="GtkComboBoxText">
         <property name="margin-top">12</property>
         <property name="margin-bottom">12</property>
         <property name="margin-start">12</property>
         <property name="margin-end">12</property>
-        <style>
-          <class name="button_1"/>
-        </style>
+        <property name="active-id">factory</property>
+        <items>
+          <item translatable="yes" id="factory">Factory</item>
+          <item translatable="yes" id="home">Home</item>
+          <item translatable="yes" id="subway">Subway</item>
+        </items>
       </object>
     </child>
   </template>
