+++ book/listings/css/8/style.css
@@ -1,3 +1,3 @@
 combobox box arrow {
-  color: red 
+  -gtk-icon-source: -gtk-icontheme('format-justify-fill-symbolic');
 }
