--- book/listings/css/9/style.css
@@ -1,3 +1,0 @@
-combobox box arrow {
-  -gtk-icon-source: -gtk-icontheme('format-justify-fill-symbolic');
-}
