--- book/listings/css/9/window/window.ui
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<interface>
-  <template class="MyGtkAppWindow" parent="GtkApplicationWindow">
-    <property name="title" translatable="yes">My GTK App</property>
-    <child>
-      <object class="GtkComboBoxText">
-        <property name="margin-top">12</property>
-        <property name="margin-bottom">12</property>
-        <property name="margin-start">12</property>
-        <property name="margin-end">12</property>
-        <property name="active-id">factory</property>
-        <items>
-          <item translatable="yes" id="factory">Factory</item>
-          <item translatable="yes" id="home">Home</item>
-          <item translatable="yes" id="subway">Subway</item>
-        </items>
-      </object>
-    </child>
-  </template>
-</interface>
